+++ CHANGELOG.md
@@ -1,4 +1,11 @@
-# master
+# 0.1.0
+
+* Updated for latest ember-cli
+* *BREAKING CHANGE* the mixin is now located at `ember-data-route`
+  instead of the previous import path of
+`ember-data-route/mixins/data-route`
+
+# 0.0.4
 
 * Update to use its own es6 namespace [#2](https://github.com/dockyard/ember-data-route/pull/2)
 * Update for ember-cli 0.0.41 [#2](https://github.com/dockyard/ember-data-route/pull/2)